Fixed variable narrowing warnings.

+++ examples/common/imgui/imgui.cpp
@@ -456,7 +456,7 @@ struct Imgui
 			}
 		}
 
-		return bgfx::createTexture2D(_width, _height, 0, bgfx::TextureFormat::BGRA8, 0, mem);
+		return bgfx::createTexture2D(uint16_t(_width), uint16_t(_height), 0, bgfx::TextureFormat::BGRA8, 0, mem);
 	}
 
 	ImguiFontHandle create(const void* _data, uint32_t _size, float _fontSize)
@@ -910,7 +910,7 @@ struct Imgui
 		Area& area = getCurrentArea();
 
 		const uint16_t parentBottom = parentArea.m_scissorY + parentArea.m_scissorHeight;
-		const uint16_t childBottom = parentArea.m_widgetY + _height;
+		const uint16_t childBottom  = parentArea.m_widgetY + _height;
 		const uint16_t bottom = IMGUI_MIN(childBottom, parentBottom);
 
 		const uint16_t top = IMGUI_MAX(parentArea.m_widgetY, parentArea.m_scissorY);
